[cri] Move instrument service to a separate Go package
